% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{createStanModel}
\alias{createStanModel}
\title{Create an .stan model}
\usage{
createStanModel(L0, P, Lp = NULL)
}
\arguments{
\item{L0}{The number of principal components kept}

\item{P}{The number of covariates}

\item{Lp}{A vector for the number of basis function for each covariate}
}
\value{
The .stan model
}
\description{
This is an internal function used to create a Stan model.
}
\examples{
cat(createStanModel(3, 5, 1:5))
}
