% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_trend.R
\name{proba_trend}
\alias{proba_trend}
\title{Calculate the probability of a trend}
\usage{
proba_trend(
  modeldata,
  model_res,
  lag,
  trend_direction,
  ID,
  date,
  value,
  dataframe.format = TRUE
)
}
\arguments{
\item{modeldata}{The long-format data frame/table of virus concentration}

\item{model_res}{The fitted model from WWmodel()}

\item{lag}{The time lag D in working days (See the manuscript)}

\item{trend_direction}{String. Direction of the trend to calculate the probability:
\code{up} or \code{down}.}

\item{dataframe.format}{Boolean. Is the result returnedas a dataframe?}
}
\value{

}
\description{
Calculate the probability of a trend
}
