% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{PROBplot}
\alias{PROBplot}
\title{Calculate the probability of increase in virus concentration}
\usage{
PROBplot(modeldata, model_res, lag, ncol)
}
\arguments{
\item{modeldata}{The long-format data frame/table of virus concentration}

\item{model_res}{The fitted model from WWmodel()}

\item{lag}{The time lag D in working days (See the manuscript)}

\item{ncol}{The number of columns in the plot}
}
\value{
The plot of probabilities.
}
\description{
This function is used to calculate the probability of increase in virus concentration. A plot is returned.
}
\examples{
PROBplot(modeldata, model_res, 5, 4)
}
