% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{SAMPLEplot}
\alias{SAMPLEplot}
\title{Create a plot of the sampled curves}
\usage{
SAMPLEplot(modeldata, model_res, ncol)
}
\arguments{
\item{modeldata}{The long-format data frame/table of virus concentration}

\item{model_res}{The fitted model from WWmodel()}

\item{ncol}{The number of columns in the plot}
}
\value{
The plot of simultaneous confidence bands.
}
\description{
This function is used to create a plot of simultaneous confidence bands.
}
\examples{
SAMPLEplot(modeldata, model_res, 4)
}
