% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\name{WWforecast}
\alias{WWforecast}
\title{Main forecasting function}
\usage{
WWforecast(
  h.ahead,
  modeldata,
  model_res,
  ID,
  date,
  value,
  covariate = NULL,
  iteration,
  burnin
)
}
\arguments{
\item{h.ahead}{The number of steps ahead for which prediction is required}

\item{modeldata}{The long-format data frame/table of virus concentration}

\item{model_res}{The model result from WWmodel()}

\item{ID}{Names of curve IDs (used to identify a unique curve)}

\item{date}{Name of date column}

\item{value}{Name of value column}

\item{covariate}{Names of covariates (default is NULL)}

\item{iteration}{A positive integer specifying the number of iterations for each chain (including burnin).}

\item{burnin}{The number of burnin iterations}
}
\value{
Ypred The forecast of observed virus concentration
}
\description{
This is the main forecasting function for a fitted WWmodel.
}
\details{
This function is used to make forecasts using a fitted WWmodel.
See ?FORECASTplot for how to plot the forecasts.
}
\examples{
rawdata = as.data.table(readRDS("ww-db-2021-09-10.rds"))
modeldata = DataPrep(rawdata, "N1")

ID = c("Location", "target", "replicate")
date = "date"
value = "log10.value.raw"
covariate = c("dinflvol", "temp")
model_res = WWmodel(modeldata, ID, date, value, covariate, 5000, 2500)

forecast_res = WWforecast(5, modeldata, model_res, ID, date, value, covariate, 5000, 2500)
}
