% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocess.R
\name{DataPrep}
\alias{DataPrep}
\title{The data preparation function}
\usage{
DataPrep(RawData, model_target)
}
\arguments{
\item{RawData}{The raw data for preprocessing}

\item{model_target}{The target virus concentration (N1 or N2)}
}
\value{
The data frame/table for the main function
}
\description{
This function is used to prepare a data frame/table for the main function.
}
\examples{
rawdata = as.data.table(readRDS("ww-db-2021-09-10.rds"))
modeldata = DataPrep(rawdata, "N1")
}
